requirements.txt
attrs==24.2.0
bech32==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cosmpy==0.9.2
ecdsa==0.19.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
protobuf==4.25.5
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
pyuno==0.1.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
urllib3==2.2.3
